requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.9
requests==2.23.0
urllib3==1.25.8
pathvalidate==2.2.0
hurry.filesize==0.9
sqlalchemy==1.3.13
